group files to gui_utility and gui_elements
